requirements.txt
annotated-types==0.6.0
ansicon==1.89.0
anyio==3.7.1
arrow==1.3.0
asgiref==3.8.1
attrs==23.1.0
autopep8==2.0.2
beautifulsoup4==4.12.2
blessed==1.20.0
blinker==1.4
bs4==0.0.1
CacheControl==0.12.10
cachy==0.3.0
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
cleo==0.8.1
clikit==0.6.2
cmake==3.24.1.1
colorama==0.4.6
command-not-found==0.3
crashtest==0.3.1
cryptography==41.0.5
dbus-python==1.2.18
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.4
distro==1.9.0
distro-info==1.1+ubuntu0.1
dj-database-url==2.2.0
Django==5.1.1
django-auto-prefetching==0.2.12
django-cors-headers==3.13.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.1
django-filter==22.1
django-heroku==0.3.1
django-picklefield==3.2
django-q==1.3.9
django-q2==1.7.2
django-silk==5.2.0
django-storages==1.14.4
djangorestframework==3.14.0
environ==1.0
exceptiongroup==1.1.2
factory-boy==3.3.0
Faker==23.3.0
filelock==3.6.0
gprof2dot==2022.7.29
gyp==0.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httplib2==0.20.2
httpx==0.26.0
idna==3.4
imageio==2.36.0
imageio-ffmpeg==0.5.1
importlib-metadata==4.6.4
inflection==0.5.1
iniconfig==2.0.0
jeepney==0.7.1
jinxed==1.3.0
keyring
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
Markdown==3.4.1
mongoengine==0.24.2
more-itertools==8.10.0
moviepy==1.0.3
msgpack==1.0.3
netifaces==0.11.0
numpy==2.1.2
oauthlib==3.2.1
openai==1.3.5
outcome==1.2.0
packaging==24.1
pastel==0.2.1
pexpect==4.8.0
Pillow==10.0.1
pipdeptree==2.23.4
pkginfo==1.8.2
platformdirs==2.5.1
pluggy==1.4.0
poetry
poetry-core
postgres==4.0
proglog==0.1.10
psycopg2==2.9.9
psycopg2-binary==2.9.6
psycopg2-pool==1.1
ptyprocess==0.7.0
pycodestyle==2.11.0
pycparser==2.21
pycryptodome==3.19.1
pycryptodome-test-vectors==1.0.13
pydantic==2.6.0
pydantic_core==2.16.1
PyGObject==3.42.1
PyJWT==2.5.0
pylev==1.2.0
pymongo==4.2.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==8.0.2
pytest-django==4.9.0
pytest-factoryboy==2.6.0
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.4
PyYAML==5.4.1
redis==3.5.3
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
SecretStorage==3.3.1
selenium==4.25.0
shellingham==1.4.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.0.0
social-auth-app-django-mongoengine==1.0.0
social-auth-core==4.3.0
social-auth-storage-mongoengine==1.0.1
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.3
style==1.1.0
systemd-python==234
tomli==2.0.1
tomlkit==0.9.2
tqdm==4.66.1
trio==0.22.1
trio-websocket==0.10.3
types-python-dateutil==2.9.0.20240821
typing_extensions==4.10.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
undetected-geckodriver==1.0.3
update==0.0.1
urllib3==1.26.12
virtualenv==20.13.0+ds
wadllib==1.3.6
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
whitenoise==6.7.0
wsproto==1.2.0
zipp==1.0.0
